src/requirements/dev.txt
-r base.txt

ipdb==0.13.3
black==19.10b0

pytest==6.0.1
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-env==0.6.2
pytest-pylint==0.17.0
pytest-pythonpath==0.7.3
pytest-sugar==0.9.4
pyhamcrest==2.0.2